emps/requirements.txt
django>2
djangorestframework==3.9.4